+++ src/main/java/agents/flexibility/dynamicProgramming/assessment/AssessmentFunctionBuilder.java
@@ -8,7 +8,9 @@ import de.dlr.gitlab.fame.agent.input.ParameterData;
 import de.dlr.gitlab.fame.agent.input.ParameterData.MissingDataException;
 import de.dlr.gitlab.fame.agent.input.Tree;
 
-/** Builds {@link AssessmentFunction}s */
+/** Builds {@link AssessmentFunction}s from provided input parameters
+ * 
+ * @author Christoph Schimeczek, Felix Nitsch, Johannes Kochems */
 public final class AssessmentFunctionBuilder {
 	public static final Tree parameters = Make.newTree().add(Make.newEnum("Type", Type.class)).buildTree();
 
@@ -16,6 +18,10 @@ public final class AssessmentFunctionBuilder {
 	enum Type {
 		/** Maximise profit using an electricity price forecast neglecting any price impact of bids */
 		MAX_PROFIT_PRICE_TAKER,
+		/** Minimise total system costs using a merit order forecast that takes into account the impact of own bids */
+		SINGLE_AGENT_MIN_SYSTEM_COST,
+		/** Maximise own profits using a merit order forecast that takes into account the impact of own bids */
+		SINGLE_AGENT_MAX_PROFIT,
 	}
 
 	public static final String ERR_NOT_IMPLEMENTED = "Assessment function is not implemented: ";
@@ -25,6 +31,10 @@ public final class AssessmentFunctionBuilder {
 		switch (type) {
 			case MAX_PROFIT_PRICE_TAKER:
 				return new MaxProfitPriceTaker();
+			case SINGLE_AGENT_MIN_SYSTEM_COST:
+				return new MinSystemCost();
+			case SINGLE_AGENT_MAX_PROFIT:
+				return new MaxProfit();
 			default:
 				throw new RuntimeException(ERR_NOT_IMPLEMENTED + type);
 		}