+++ src/main/java/agents/markets/meritOrder/sensitivities/MeritOrderSensitivity.java
@@ -26,6 +26,11 @@ public abstract class MeritOrderSensitivity {
 	 * associated flexibility device */
 	protected ArrayList<SensitivityItem> dischargingItems = new ArrayList<>();
 
+	int lastChargeIndex = 0;
+	int lastDischargeIndex = 0;
+	double lastChargeEnergy = 0.;
+	double lastDischargeEnergy = 0.;
+
 	/** sets maximum charging and discharging powers in MW according to specified values
 	 * 
 	 * @param maxChargePowerInMW to be set
@@ -46,8 +51,6 @@ public abstract class MeritOrderSensitivity {
 		dischargingItems.sort(getComparator().reversed());
 		setCumulativeValues(chargingItems);
 		setCumulativeValues(dischargingItems);
-		chargingItems.removeIf(i -> i.getCumulatedLowerPower() > externalChargingPowerInMW);
-		dischargingItems.removeIf(i -> i.getCumulatedLowerPower() > externalDischargingPowerInMW);
 	}
 
 	/** clears stored sensitivity data */
@@ -126,6 +129,55 @@ public abstract class MeritOrderSensitivity {
 		}
 	}
 
+	/** Return value for given external energy
+	 * 
+	 * @param externalEnergyDeltaInMWH &gt;0: charging, &lt;0: discharging
+	 * @return value */
+	public double getValue(double externalEnergyDeltaInMWH) {
+		if (externalEnergyDeltaInMWH > 0) {
+			return getValueCharging(externalEnergyDeltaInMWH);
+		} else if (externalEnergyDeltaInMWH < 0) {
+			return getValueDischarging(externalEnergyDeltaInMWH);
+		}
+		return 0;
+	}
+
+	/** @return value associated with the specified energy delta when charging */
+	private double getValueCharging(double externalEnergyDeltaInMWH) {
+		int firstIndex = externalEnergyDeltaInMWH >= lastChargeEnergy ? lastChargeIndex : 0;
+		for (int index = firstIndex; index < chargingItems.size(); index++) {
+			var item = chargingItems.get(index);
+			if (item.getCumulatedUpperPower() >= externalEnergyDeltaInMWH) {
+				lastChargeEnergy = externalEnergyDeltaInMWH;
+				lastChargeIndex = index;
+				return calcValueOfItemAtPower(item, externalEnergyDeltaInMWH);
+			}
+		}
+		return Double.NaN;
+	}
+
+	/** @return value associated with the specified energy delta when discharging */
+	private double getValueDischarging(double externalEnergyDeltaInMWH) {
+		int firstIndex = externalEnergyDeltaInMWH <= lastDischargeEnergy ? lastDischargeIndex : 0;
+		for (int index = firstIndex; index < dischargingItems.size(); index++) {
+			var item = dischargingItems.get(index);
+			if (item.getCumulatedUpperPower() >= -externalEnergyDeltaInMWH) {
+				lastDischargeEnergy = externalEnergyDeltaInMWH;
+				lastDischargeIndex = index;
+				return calcValueOfItemAtPower(item, externalEnergyDeltaInMWH);
+			}
+		}
+		return Double.NaN;
+	}
+
+	/** Calculates value of given sensitivity item at specified power
+	 * 
+	 * @param item that is to be evaluated
+	 * @param power to be applied
+	 * @return value of given {@link SensitivityItem} at specified power according to this Sensitivity type;<br>
+	 *         when power &gt; 0: <b>charging</b>, otherwise <b>discharging</b> */
+	protected abstract double calcValueOfItemAtPower(SensitivityItem item, double power);
+
 	/** Calculates monetary value of given item
 	 * 
 	 * @param item to assess
@@ -139,40 +191,22 @@ public abstract class MeritOrderSensitivity {
 	 *         powers; first entry corresponds to maximum discharging power, while the last entry resembles sensitivity value at
 	 *         maximum charging power */
 	public double[] getValuesInSteps(int numberOfTransitionSteps) {
+		chargingItems.removeIf(i -> i.getCumulatedLowerPower() > externalChargingPowerInMW);
+		dischargingItems.removeIf(i -> i.getCumulatedLowerPower() > externalDischargingPowerInMW);
+
 		double[] values = new double[2 * numberOfTransitionSteps + 1];
+		values[numberOfTransitionSteps] = 0.0;
 
 		double chargingPowerPerStep = externalChargingPowerInMW / numberOfTransitionSteps;
-		int maxChargingIndex = chargingItems.size();
-		int chargingIndex = 0;
-		values[numberOfTransitionSteps] = 0.0;
 		for (int step = 1; step <= numberOfTransitionSteps; step++) {
 			double power = chargingPowerPerStep * step;
-			while (chargingIndex < maxChargingIndex && chargingItems.get(chargingIndex).getCumulatedUpperPower() < power) {
-				chargingIndex++;
-			}
-			int indexInArray = numberOfTransitionSteps + step;
-			if (chargingIndex < maxChargingIndex) {
-				values[indexInArray] = calcValueOfItemAtPower(chargingItems.get(chargingIndex), power);
-			} else {
-				values[indexInArray] = Double.NaN;
-			}
+			values[numberOfTransitionSteps + step] = getValueCharging(power);
 		}
 
-		double dischargingPowerPerStep = externalDischargingPowerInMW / numberOfTransitionSteps;
-		int dischargingIndex = 0;
-		int maxDischargingIndex = dischargingItems.size();
+		double dischargingPowerPerStep = -externalDischargingPowerInMW / numberOfTransitionSteps;
 		for (int step = 1; step <= numberOfTransitionSteps; step++) {
 			double power = dischargingPowerPerStep * step;
-			while (dischargingIndex < maxDischargingIndex
-					&& dischargingItems.get(dischargingIndex).getCumulatedUpperPower() < power) {
-				dischargingIndex++;
-			}
-			int indexInArray = numberOfTransitionSteps - step;
-			if (dischargingIndex < maxDischargingIndex) {
-				values[indexInArray] = calcValueOfItemAtPower(dischargingItems.get(dischargingIndex), -power);
-			} else {
-				values[indexInArray] = Double.NaN;
-			}
+			values[numberOfTransitionSteps - step] = getValueDischarging(power);
 		}
 		return values;
 	}
@@ -184,14 +218,6 @@ public abstract class MeritOrderSensitivity {
 	public final StepPower getStepPowers(int numberOfTransitionSteps) {
 		return new StepPower(externalChargingPowerInMW, externalDischargingPowerInMW, numberOfTransitionSteps);
 	}
-
-	/** Calculates value of given sensitivity item at specified power
-	 * 
-	 * @param item that is to be evaluated
-	 * @param power to be applied
-	 * @return value of given {@link SensitivityItem} at specified power according to this Sensitivity type;<br>
-	 *         when power &gt; 0: <b>charging</b>, otherwise <b>discharging</b> */
-	protected abstract double calcValueOfItemAtPower(SensitivityItem item, double power);
 
 	/** Stores fixed electricity price forecast value
 	 *